feat(sdk): move in-scope /v1 families into typed SDK clients
- projects-client: billing (account-state), channels, gateway, apps, transcription
- platform-client: instance-admin (sandbox admin + provision-stream URL builder)
- explicit-token server variants for Next server actions (no configureKortix singleton in per-request contexts)
- web modules reduced to thin re-exports; consumers unchanged
