+++ apps/web/src/hooks/channels/use-channels-installations.ts
@@ -1,14 +1,25 @@
 'use client';
 
-import { backendApi } from '@/lib/api-client';
+import {
+  connectEmail,
+  connectSlack,
+  disconnectEmail,
+  disconnectSlack,
+  getEmailInstallation,
+  getEmailMode,
+  getSlackInstallation,
+  getSlackManifest,
+  getSlackMode,
+  updateEmailPolicy,
+  type EmailInstallation,
+  type EmailMode,
+  type EmailSenderPolicy,
+  type SlackInstallation,
+  type SlackMode,
+} from '@kortix/sdk/projects-client';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
-export interface SlackInstallation {
-  workspaceId: string;
-  workspaceName: string | null;
-  botUserId: string | null;
-  installedAt: string;
-}
+export type { EmailInstallation, EmailMode, EmailSenderPolicy, SlackInstallation, SlackMode };
 
 const key = (projectId: string | null) =>
   ['channels', 'slack-install', projectId ?? 'none'] as const;
@@ -18,15 +29,7 @@ export function useSlackInstall(projectId: string | null) {
     queryKey: key(projectId),
     enabled: !!projectId,
     staleTime: 30_000,
-    queryFn: async () => {
-      if (!projectId) return null;
-      const res = await backendApi.get<SlackInstallation | null>(
-        `/projects/${encodeURIComponent(projectId)}/channels/slack/installation`,
-        { showErrors: false },
-      );
-      if (!res.success) return null;
-      return res.data ?? null;
-    },
+    queryFn: () => (projectId ? getSlackInstallation(projectId) : null),
   });
 }
 
@@ -39,27 +42,12 @@ interface ConnectInput {
 export function useConnectSlack() {
   const qc = useQueryClient();
   return useMutation({
-    mutationFn: async ({ projectId, ...body }: ConnectInput) => {
-      const res = await backendApi.post<SlackInstallation>(
-        `/projects/${encodeURIComponent(projectId)}/channels/slack/connect`,
-        body,
-        { showErrors: false },
-      );
-      if (!res.success || !res.data) {
-        throw new Error(res.error?.message ?? 'Failed to connect');
-      }
-      return res.data;
-    },
+    mutationFn: ({ projectId, ...body }: ConnectInput) => connectSlack(projectId, body),
     onSuccess: (_data, { projectId }) => {
       qc.invalidateQueries({ queryKey: key(projectId) });
       qc.invalidateQueries({ queryKey: ['project-connectors', projectId] });
     },
   });
-}
-
-export interface SlackMode {
-  oauth_available: boolean;
-  install_url: string | null;
 }
 
 const modeKey = (projectId: string | null) =>
@@ -70,17 +58,8 @@ export function useSlackMode(projectId: string | null) {
     queryKey: modeKey(projectId),
     enabled: !!projectId,
     staleTime: 60_000,
-    queryFn: async () => {
-      if (!projectId) return { oauth_available: false, install_url: null } satisfies SlackMode;
-      const res = await backendApi.get<SlackMode>(
-        `/projects/${encodeURIComponent(projectId)}/channels/slack/mode`,
-        { showErrors: false },
-      );
-      if (!res.success || !res.data) {
-        return { oauth_available: false, install_url: null } satisfies SlackMode;
-      }
-      return res.data;
-    },
+    queryFn: () =>
+      projectId ? getSlackMode(projectId) : ({ oauth_available: false, install_url: null } satisfies SlackMode),
   });
 }
 
@@ -94,14 +73,8 @@ export function useSlackManifest(projectId: string | null) {
     staleTime: 60_000,
     queryFn: async () => {
       if (!projectId) return null;
-      const res = await backendApi.get<Record<string, unknown>>(
-        `/webhooks/slack/${encodeURIComponent(projectId)}/manifest`,
-        { showErrors: false },
-      );
-      if (!res.success || !res.data) {
-        throw new Error(res.error?.message ?? 'Failed to load Slack manifest');
-      }
-      return JSON.stringify(res.data, null, 2);
+      const manifest = await getSlackManifest(projectId);
+      return JSON.stringify(manifest, null, 2);
     },
   });
 }
@@ -109,41 +82,12 @@ export function useSlackManifest(projectId: string | null) {
 export function useDisconnectSlack() {
   const qc = useQueryClient();
   return useMutation({
-    mutationFn: async (projectId: string) => {
-      const res = await backendApi.delete(
-        `/projects/${encodeURIComponent(projectId)}/channels/slack/installation`,
-        { showErrors: false },
-      );
-      if (!res.success) throw new Error(res.error?.message ?? 'Failed to disconnect');
-    },
+    mutationFn: (projectId: string) => disconnectSlack(projectId),
     onSuccess: (_data, projectId) => {
       qc.invalidateQueries({ queryKey: key(projectId) });
       qc.invalidateQueries({ queryKey: ['project-connectors', projectId] });
     },
   });
-}
-
-export interface EmailInstallation {
-  profileSlug: string;
-  inboxId: string;
-  email: string;
-  displayName: string | null;
-  webhookId: string | null;
-  senderPolicy: EmailSenderPolicy;
-  installedAt: string;
-}
-
-export interface EmailSenderPolicy {
-  mode: 'allow_all' | 'restricted';
-  allowedEmails: string[];
-  allowedDomains: string[];
-  allowedRegex: string | null;
-}
-
-export interface EmailMode {
-  provider: 'agentmail';
-  enabled?: boolean;
-  managed_available: boolean;
 }
 
 const emailKey = (projectId: string | null, connectorSlug?: string | null) =>
@@ -156,16 +100,7 @@ export function useEmailInstall(projectId: string | null, connectorSlug?: string
     queryKey: emailKey(projectId, connectorSlug),
     enabled: !!projectId,
     staleTime: 30_000,
-    queryFn: async () => {
-      if (!projectId) return null;
-      const query = connectorSlug ? `?connector_slug=${encodeURIComponent(connectorSlug)}` : '';
-      const res = await backendApi.get<EmailInstallation | null>(
-        `/projects/${encodeURIComponent(projectId)}/channels/email/installation${query}`,
-        { showErrors: false },
-      );
-      if (!res.success) return null;
-      return res.data ?? null;
-    },
+    queryFn: () => (projectId ? getEmailInstallation(projectId, connectorSlug) : null),
   });
 }
 
@@ -174,19 +109,10 @@ export function useEmailMode(projectId: string | null) {
     queryKey: emailModeKey(projectId),
     enabled: !!projectId,
     staleTime: 60_000,
-    queryFn: async () => {
-      if (!projectId) {
-        return { provider: 'agentmail', managed_available: false } satisfies EmailMode;
-      }
-      const res = await backendApi.get<EmailMode>(
-        `/projects/${encodeURIComponent(projectId)}/channels/email/mode`,
-        { showErrors: false },
-      );
-      if (!res.success || !res.data) {
-        return { provider: 'agentmail', managed_available: false } satisfies EmailMode;
-      }
-      return res.data;
-    },
+    queryFn: () =>
+      projectId
+        ? getEmailMode(projectId)
+        : ({ provider: 'agentmail', managed_available: false } satisfies EmailMode),
   });
 }
 
@@ -205,17 +131,7 @@ interface ConnectEmailInput {
 export function useConnectEmail() {
   const qc = useQueryClient();
   return useMutation({
-    mutationFn: async ({ projectId, ...body }: ConnectEmailInput) => {
-      const res = await backendApi.post<EmailInstallation>(
-        `/projects/${encodeURIComponent(projectId)}/channels/email/connect`,
-        body,
-        { showErrors: false },
-      );
-      if (!res.success || !res.data) {
-        throw new Error(res.error?.message ?? 'Failed to connect email');
-      }
-      return res.data;
-    },
+    mutationFn: ({ projectId, ...body }: ConnectEmailInput) => connectEmail(projectId, body),
     onSuccess: (_data, { projectId, connector_slug }) => {
       qc.invalidateQueries({ queryKey: emailKey(projectId) });
       qc.invalidateQueries({ queryKey: emailKey(projectId, connector_slug) });
@@ -230,12 +146,7 @@ export function useDisconnectEmail() {
     mutationFn: async (input: string | { projectId: string; connectorSlug?: string | null }) => {
       const projectId = typeof input === 'string' ? input : input.projectId;
       const connectorSlug = typeof input === 'string' ? null : input.connectorSlug;
-      const query = connectorSlug ? `?connector_slug=${encodeURIComponent(connectorSlug)}` : '';
-      const res = await backendApi.delete(
-        `/projects/${encodeURIComponent(projectId)}/channels/email/installation${query}`,
-        { showErrors: false },
-      );
-      if (!res.success) throw new Error(res.error?.message ?? 'Failed to disconnect email');
+      await disconnectEmail(projectId, connectorSlug);
       return { projectId, connectorSlug };
     },
     onSuccess: ({ projectId, connectorSlug }) => {
@@ -248,7 +159,7 @@ export function useDisconnectEmail() {
 export function useUpdateEmailPolicy() {
   const qc = useQueryClient();
   return useMutation({
-    mutationFn: async ({
+    mutationFn: ({
       projectId,
       connectorSlug,
       sender_policy,
@@ -256,16 +167,7 @@ export function useUpdateEmailPolicy() {
       projectId: string;
       connectorSlug?: string | null;
       sender_policy: EmailSenderPolicy;
-    }) => {
-      const res = await backendApi.patch<EmailInstallation>(
-        `/projects/${encodeURIComponent(projectId)}/channels/email/installation`,
-        { connector_slug: connectorSlug ?? 'kortix_email', sender_policy },
-        { showErrors: false },
-      );
-      if (!res.success || !res.data)
-        throw new Error(res.error?.message ?? 'Failed to update email policy');
-      return res.data;
-    },
+    }) => updateEmailPolicy(projectId, connectorSlug, sender_policy),
     onSuccess: (_data, { projectId, connectorSlug }) => {
       qc.invalidateQueries({ queryKey: emailKey(projectId, connectorSlug) });
       qc.invalidateQueries({ queryKey: ['project-connectors', projectId] });
