+++ apps/web/src/lib/api/billing.ts
@@ -1,162 +1,20 @@
 import { backendApi } from "../api-client";
+import {
+  getAccountState as sdkGetAccountState,
+  type AccountState,
+} from '@kortix/sdk/projects-client';
 
 // =============================================================================
 // UNIFIED ACCOUNT STATE - Primary API for all billing data
+//
+// The account-state type + read now live in the SDK
+// (packages/sdk/src/platform/projects-client/billing.ts) — it's the
+// entitlement-gating data, not product-UI. Re-exported here so existing
+// `import { AccountState } from '@/lib/api/billing'` call sites keep working.
+// The Stripe checkout/portal/credits/auto-topup mutations below stay web-local.
 // =============================================================================
 
-export interface AccountState {
-  credits: {
-    total: number;
-    daily: number;
-    monthly: number;
-    extra: number;
-    can_run: boolean;
-    daily_refresh: {
-      enabled: boolean;
-      daily_amount: number;
-      refresh_interval_hours: number;
-      last_refresh?: string;
-      next_refresh_at?: string;
-      seconds_until_refresh?: number;
-    } | null;
-  };
-  subscription: {
-    tier_key: string;
-    tier_display_name: string;
-    status: string;
-    billing_period: 'monthly' | 'yearly' | 'yearly_commitment' | null;
-    provider: 'stripe' | 'revenuecat' | 'local';
-    subscription_id: string | null;
-    current_period_end: number | null;
-    cancel_at_period_end: boolean;
-    is_cancelled: boolean;
-    cancellation_effective_date: string | null;
-    has_scheduled_change: boolean;
-    scheduled_change: {
-      type: 'downgrade';
-      current_tier: {
-        name: string;
-        display_name: string;
-        monthly_credits?: number;
-      };
-      target_tier: {
-        name: string;
-        display_name: string;
-        monthly_credits?: number;
-      };
-      effective_date: string;
-    } | null;
-    commitment: {
-      has_commitment: boolean;
-      can_cancel: boolean;
-      commitment_type?: string | null;
-      months_remaining?: number | null;
-      commitment_end_date?: string | null;
-    };
-    can_purchase_credits: boolean;
-  };
-  models: Array<{
-    id: string;
-    name: string;
-    provider: string;
-    allowed: boolean;
-    context_window: number;
-    capabilities: string[];
-    priority: number;
-  }>;
-  limits?: {
-    concurrent_runs: {
-      running_count: number;
-      limit: number;
-      can_start: boolean;
-      tier_name: string;
-    };
-    ai_worker_count: {
-      current_count: number;
-      limit: number;
-      can_create: boolean;
-      tier_name: string;
-    };
-    custom_mcp_count: {
-      current_count: number;
-      limit: number;
-      can_create: boolean;
-      tier_name: string;
-    };
-    concurrent_sessions?: {
-      active: number;
-      limit: number;
-    };
-  };
-  tier: {
-    name: string;
-    display_name: string;
-    monthly_credits: number;
-    can_purchase_credits: boolean;
-    /** Enterprise feature gates for this tier (SSO / SCIM). Drives whether the
-     *  account-settings "Identity & directory" cards render. */
-    entitlements?: {
-      sso: boolean;
-      scim: boolean;
-    };
-  };
-  auto_topup?: {
-    enabled: boolean;
-    threshold: number;
-    amount: number;
-  };
-  instances?: Array<{
-    sandbox_id: string;
-    name: string;
-    provider: string;
-    status: string;
-    server_type: string | null;
-    location: string | null;
-    is_included: boolean;
-    stripe_subscription_id: string | null;
-    stripe_subscription_item_id: string | null;
-    cancel_at_period_end?: boolean;
-    cancel_at?: string | null;
-    created_at: string;
-  }>;
-  can_add_instances?: boolean;
-  can_claim_computer?: boolean;
-  // Whether the CURRENT user may change billing for this account (billing.write —
-  // owners only by default). Drives the "Subscribe" / "Manage billing" CTA gate:
-  // members (billing.read only) see a disabled CTA instead of clicking through to
-  // a 403. UI hint only — the billing API enforces the same gate server-side.
-  // Absent (undefined) is treated as "allowed" so older responses don't block owners.
-  can_manage_billing?: boolean;
-  // True only for genuine legacy per-machine accounts that have a machine to
-  // migrate — gates the "Claim seat-based pricing" card (new per-seat-era users
-  // must not see it, or the claim dead-ends on "nothing to switch").
-  can_claim_per_seat?: boolean;
-  // Billing v2 — present for accounts on the new per-seat plan.
-  billing_model?: 'legacy' | 'per_seat';
-  seats?: {
-    count: number;
-    price_per_seat_usd: number;
-    typical_compute_budget_per_seat_usd: number;
-    typical_llm_budget_per_seat_usd: number;
-  };
-  // Live account-member count = the seat quantity a per-seat subscribe is billed
-  // for right now (server uses the same count for the Stripe line item). Lets the
-  // subscribe modal show the real projected total before redirecting to Stripe.
-  // Present once the API is updated; absent → fall back to seats.count then 1.
-  member_count?: number;
-  usage_this_period?: {
-    compute_usd: number;
-    llm_usd: number;
-    total_usd: number;
-    period_start: string | null;
-    period_end: string | null;
-  } | null;
-  _cache?: {
-    cached: boolean;
-    ttl_seconds?: number;
-    local_mode?: boolean;
-  };
-}
+export type { AccountState };
 
 // =============================================================================
 // MUTATION REQUEST/RESPONSE TYPES
@@ -325,68 +183,6 @@ export interface CheckoutSessionDetails {
   payment_status: string;
 }
 
-function getDefaultAccountState(): AccountState {
-  return {
-    credits: {
-      total: 0,
-      daily: 0,
-      monthly: 0,
-      extra: 0,
-      can_run: false,
-      daily_refresh: null,
-    },
-    subscription: {
-      tier_key: 'none',
-      tier_display_name: 'No Plan',
-      status: 'no_subscription',
-      billing_period: null,
-      provider: 'stripe',
-      subscription_id: null,
-      current_period_end: null,
-      cancel_at_period_end: false,
-      is_cancelled: false,
-      cancellation_effective_date: null,
-      has_scheduled_change: false,
-      scheduled_change: null,
-      commitment: {
-        has_commitment: false,
-        can_cancel: true,
-        commitment_type: null,
-        months_remaining: null,
-        commitment_end_date: null,
-      },
-      can_purchase_credits: false,
-    },
-    models: [],
-    limits: {
-      concurrent_runs: {
-        running_count: 0,
-        limit: 0,
-        can_start: false,
-        tier_name: 'none'
-      },
-      ai_worker_count: {
-        current_count: 0,
-        limit: 0,
-        can_create: false,
-        tier_name: 'none'
-      },
-      custom_mcp_count: {
-        current_count: 0,
-        limit: 0,
-        can_create: false,
-        tier_name: 'none'
-      },
-    },
-    tier: {
-      name: 'none',
-      display_name: 'No Plan',
-      monthly_credits: 0,
-      can_purchase_credits: false,
-    },
-  };
-}
-
 // =============================================================================
 // BILLING API
 // =============================================================================
@@ -400,25 +196,14 @@ export const billingApi = {
    * member of (e.g. on /accounts/[id] pages). Without it, the backend uses the
    * user's first membership — fine for global surfaces like the user menu and
    * upgrade dialog, but wrong for the per-account billing page.
+   *
+   * The actual fetch + graceful-degrade logic lives in the SDK
+   * (`@kortix/sdk/projects-client`'s `getAccountState`) — this is a thin
+   * pass-through so existing `billingApi.getAccountState(...)` call sites
+   * don't need to change.
    */
   async getAccountState(skipCache = false, accountId?: string): Promise<AccountState> {
-    const search = new URLSearchParams();
-    if (skipCache) search.set('skip_cache', 'true');
-    if (accountId) search.set('account_id', accountId);
-    const query = search.toString();
-    const params = query ? `?${query}` : '';
-    const response = await backendApi.get<AccountState>(`/billing/account-state${params}`, {
-      showErrors: false,
-    });
-    const isGracefulDisabledResponse = response.error?.status === 404
-      && /billing is not enabled/i.test(response.error.message || '');
-    if (response.error && response.error.status !== 401 && !isGracefulDisabledResponse) {
-      throw response.error;
-    }
-    if (response.error) {
-      return getDefaultAccountState();
-    }
-    return response.data!;
+    return sdkGetAccountState({ skipCache, accountId });
   },
 
   async deductUsage(params: { amount: number; description?: string }) {
@@ -734,73 +519,20 @@ export async function getAutoTopupSetupStatus(accountId?: string): Promise<AutoT
 }
 
 // =============================================================================
-// INSTANCES (server types)
+// INSTANCES (server types) — thin re-export
+//
+// The managed-VPS ("JustAVPS") instance admin fns now live in the SDK
+// (packages/sdk/src/platform/platform-client/instance-admin.ts). Re-exported
+// here, under their original names, so existing call sites don't change.
 // =============================================================================
 
-export interface ServerType {
-  name: string;
-  description: string;
-  cores: number;
-  memory: number;
-  disk: number;
-  cpuType: 'shared' | 'dedicated';
-  architecture: 'x86' | 'arm';
-  priceMonthly: number;
-  priceMonthlyMarkup: number;
-  location: string;
-}
-
-export interface ServerTypesResponse {
-  serverTypes: ServerType[];
-  location: string;
-  defaultServerType?: string;
-  defaultLocation?: string;
-}
-
-export async function getServerTypes(location?: string): Promise<ServerTypesResponse> {
-  const params = location ? `?location=${location}` : '';
-  const response = await backendApi.get<ServerTypesResponse>(
-    `/platform/sandbox/justavps/server-types${params}`
-  );
-  if (response.error) {
-    if (response.error.status === 404 && /justavps provider is not enabled/i.test(response.error.message || '')) {
-      return {
-        serverTypes: [],
-        location: location || 'hel1',
-      };
-    }
-    throw response.error;
-  }
-  return response.data!;
-}
-
-export interface CreateInstanceRequest {
-  provider: 'justavps';
-  serverType?: string;
-  location?: string;
-  name?: string;
-  backgroundProvisioning?: boolean;
-}
-
-export async function createInstance(request: CreateInstanceRequest): Promise<any> {
-  const response = await backendApi.post<any>('/platform/sandbox', request, { timeout: 180000 });
-  if (response.error) throw response.error;
-  return response.data!;
-}
-
-export async function deleteInstance(sandboxId: string): Promise<{ success: boolean }> {
-  const response = await backendApi.delete<{ success: boolean }>(`/platform/sandbox?sandbox_id=${sandboxId}`);
-  if (response.error) throw response.error;
-  return response.data!;
-}
-
-export async function markInstanceError(sandboxId: string, errorMessage: string): Promise<void> {
-  await backendApi.post('/platform/sandbox/mark-error', { sandbox_id: sandboxId, error_message: errorMessage }, { showErrors: false, timeout: 10000 });
-}
-
-/** Claim a free default computer for legacy paid users. */
-export async function claimComputer(): Promise<any> {
-  const response = await backendApi.post<any>('/platform/sandbox/claim-computer', {}, { timeout: 60000 });
-  if (response.error) throw response.error;
-  return response.data!;
-}
+export {
+  getJustavpsServerTypes as getServerTypes,
+  createInstance,
+  deleteInstance,
+  markInstanceError,
+  claimComputer,
+  type ServerType,
+  type ServerTypesResponse,
+  type CreateInstanceRequest,
+} from '@kortix/sdk/platform-client';
