+++ apps/web/src/lib/api/transcription.ts
@@ -1,32 +1,17 @@
 import { handleApiError } from '../error-handler';
-import { backendApi } from '../api-client';
+import { transcribeAudio as sdkTranscribeAudio, type TranscriptionResponse } from '@kortix/sdk/projects-client';
 
-export interface TranscriptionResponse {
-  text: string;
-}
+export type { TranscriptionResponse };
 
+// The actual upload + error-shape lives in the SDK
+// (packages/sdk/src/platform/projects-client/transcription.ts). This wrapper
+// stays web-local only to route failures through web's `handleApiError` toast.
 export const transcribeAudio = async (audioFile: File): Promise<TranscriptionResponse> => {
   try {
-    const formData = new FormData();
-    formData.append('audio_file', audioFile);
-
-    const response = await backendApi.upload<TranscriptionResponse>(
-      '/transcription',
-      formData,
-      { showErrors: true }
-    );
-
-    if (response.error) {
-      throw new Error(
-        `Error transcribing audio: ${response.error.message} (${response.error.status})`,
-      );
-    }
-
-    return response.data!;
+    return await sdkTranscribeAudio(audioFile);
   } catch (error) {
     console.error('Failed to transcribe audio:', error);
     handleApiError(error, { operation: 'transcribe audio', resource: 'speech-to-text' });
     throw error;
   }
 };
-
