+++ packages/sdk/src/platform/platform-client/index.ts
@@ -18,3 +18,4 @@ export * from './invites';
 export * from './backups';
 export * from './ssh';
 export * from './updates';
+export * from './instance-admin';
