+++ packages/sdk/src/platform/projects-client/projects.ts
@@ -2,10 +2,13 @@
 
 import { backendApi, type ApiClientOptions } from '../api-client';
 import {
+  normalizeServerBackendBase,
+  serverTokenGet,
   unwrap,
-  type ProjectRole,
-  type ProjectGitConnection,
   type ProjectFileEntry,
+  type ProjectGitConnection,
+  type ProjectRole,
+  type ServerTokenOptions,
 } from './shared';
 
 /** Stable ids for experimental features (mirrors apps/api experimental/features). */
@@ -295,4 +298,63 @@ export async function archiveProject(projectId: string) {
   return unwrap(
     await backendApi.delete<{ ok: boolean }>(`/projects/${projectId}`),
   );
+}
+
+// ── Server-side explicit-token variants ──────────────────────────────────────
+// Next.js server actions / route handlers (post-signup first-project
+// bootstrap) run per-request with an already-resolved Supabase access token —
+// they must not rely on the SDK's process-wide `configureKortix()` seam.
+
+/**
+ * Server-side / explicit-token variant of {@link listProjectsForAccount}.
+ * Returns `null` on any failure.
+ */
+export async function fetchProjectsForAccountWithToken(
+  opts: ServerTokenOptions,
+  accountId: string,
+): Promise<KortixProject[] | null> {
+  return serverTokenGet<KortixProject[]>(
+    opts,
+    `/v1/projects?account_id=${encodeURIComponent(accountId)}`,
+  );
+}
+
+export type ProvisionProjectWithTokenResult =
+  | { ok: true; project: KortixProject }
+  | { ok: false; limitReached: boolean };
+
+/**
+ * Server-side / explicit-token variant of {@link provisionProject}. Mirrors
+ * the original bootstrap behavior: a 403 with `code: 'project_limit_reached'`
+ * is reported distinctly so the caller can fall back to re-listing existing
+ * projects instead of treating it as a hard failure.
+ */
+export async function provisionProjectWithToken(
+  opts: ServerTokenOptions,
+  input: ProvisionProjectInput,
+): Promise<ProvisionProjectWithTokenResult> {
+  if (!opts.backendUrl || !opts.accessToken) return { ok: false, limitReached: false };
+  const base = normalizeServerBackendBase(opts.backendUrl);
+  try {
+    const res = await fetch(`${base}/v1/projects/provision`, {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${opts.accessToken}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ seed_starter: true, ...input }),
+      signal: AbortSignal.timeout(opts.timeoutMs ?? 90_000),
+    });
+    if (res.ok) {
+      const project = (await res.json()) as KortixProject;
+      return { ok: true, project };
+    }
+    if (res.status === 403) {
+      const body = (await res.json().catch(() => null)) as { code?: string } | null;
+      return { ok: false, limitReached: body?.code === 'project_limit_reached' };
+    }
+    return { ok: false, limitReached: false };
+  } catch {
+    return { ok: false, limitReached: false };
+  }
 }
